etl/requirements.txt
clickhouse-driver==0.2.2
kafka-python==2.0.2
backoff==1.11.1
python-decouple==3.5
